+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # CLIFS Changelog
 
-## Unreleased 
+## Unreleased
 
+- Some refactoring:
+  - use Counter class for all counters
+  - use Enum class for storage of color constants
+  - some other minor refactors
 
 ## 0.4.2 - Oct. 13, 2023